--- app/Http/Controllers/AdminController.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-
-class AdminController extends Controller
-{
-    function index()
-    {
-        echo 'Halo, selamat datang';
-        echo '<h1>'. Auth::user()->name .'</h1>';
-        echo '<a href="/logout">Logout >></a>';
-    }
-
-    function admin()
-    {
-        echo 'Halo, selamat datang di halaman admin';
-        echo '<h1>'. Auth::user()->name .'</h1>';
-        echo '<a href="/logout">Logout >></a>';
-    }
-
-    function user()
-    {
-        echo 'Halo, selamat datang di halaman user';
-        echo '<h1>'. Auth::user()->name .'</h1>';
-        echo '<a href="/logout">Logout >></a>';
-    }
-}
