--- app/Http/Controllers/ForgotPasswordController.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Password;
-
-class ForgotPasswordController extends Controller
-{
-    function index()
-    {
-        return view('auth2.forgot-password');
-    }
-
-    function forgotPasswordHandler(Request $request)
-    {
-        $request->validate(['email' => 'required|email']);
-
-        $status = Password::sendResetLink(
-            $request->only('email')
-        );
-
-        return $status === Password::RESET_LINK_SENT
-            ? back()->with(['status' => __($status)])
-            : back()->withErrors(['email' => __($status)]);
-    }
-}
