--- resources/views/auth2/reset-password.blade.php
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Purrfect Adopt | Reset Password</title>
-    <!-- Tambahkan link ke Tailwind CSS CDN -->
-    <link href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.15/dist/tailwind.min.css" rel="stylesheet">
-</head>
-<body class="bg-gray-100">
-    <div class="flex h-screen items-center justify-center">
-        <div class="bg-white p-8 rounded shadow-md w-96">
-            <h2 class="text-2xl font-semibold mb-6">Reset Password</h2>
-
-            @include('auth2.partials.error')
-            
-            <form action="{{ route('password.update') }}" method="POST">
-                @csrf
-                <input type="hidden" name="token" value="{{ $token }}">
-                <input type="hidden" name="email" value="{{ request()->email }}">
-                <div class="mb-4">
-                    <label for="password" class="block text-gray-700 text-sm font-bold mb-2">Password:</label>
-                    <input type="password" id="password" name="password" class="w-full px-3 py-2 border border-gray-300 rounded focus:outline-none focus:border-indigo-500" placeholder="New password" required>
-                </div>
-                <div class="mb-4">
-                    <label for="password_confirmation" class="block text-gray-700 text-sm font-bold mb-2">Confirm Password:</label>
-                    <input type="password" id="password_confirmation" name="password_confirmation" class="w-full px-3 py-2 border border-gray-300 rounded focus:outline-none focus:border-indigo-500" placeholder="Confirm new password" required>
-                </div>
-                <div>
-                    <button type="submit" class="bg-indigo-500 text-white py-2 px-4 rounded hover:bg-indigo-600">Reset Password</button>
-                </div>
-            </form>
-        </div>
-    </div>
-</body>
-</html>
