+++ resources/views/components/banner.blade.php
