+++ resources/views/components/error.blade.php
