+++ resources/views/components/head.blade.php
